refactor : 입력시 dto 추가
